requirements.txt
streamlit==1.40.0
pyyaml==6.0.2
pydantic==2.9.2
jinja2==3.1.4
pandas==2.2.2
pytest==8.3.3
pytest-cov==6.0.0
plotly==5.24.1
altair==5.4.1
openai==1.54.3
httpx==0.27.2
jsonschema==4.23.0
